Overhaul coding styles

+++ src/util/usage.cpp
@@ -12,6 +12,8 @@
 #include <sys/times.h>
 #include <unistd.h>
 
+#include <gsl/narrow>
+
 using namespace dvlab::utils;
 
 void Usage::reset() {
@@ -38,7 +40,7 @@ double Usage::_check_memory() const {
 #ifdef __APPLE__
         return usage.ru_maxrss / double(1 << 20);  // bytes // NOLINT(cppcoreguidelines-pro-type-union-access) : conform to POSIX
 #else
-        return usage.ru_maxrss / double(1 << 10);  // KBytes // NOLINT(cppcoreguidelines-pro-type-union-access) : conform to POSIX
+        return gsl::narrow_cast<double>(usage.ru_maxrss) / double(1 << 10);  // KBytes // NOLINT(cppcoreguidelines-pro-type-union-access) : conform to POSIX
 #endif
     else
         return 0;
@@ -46,7 +48,7 @@ double Usage::_check_memory() const {
 double Usage::_check_tick() const {
     tms buffer{};
     times(&buffer);
-    return buffer.tms_utime;
+    return gsl::narrow_cast<double>(buffer.tms_utime);
 }
 void Usage::_set_memory_usage() { _current_memory = _check_memory() - _initial_memory; }
 void Usage::_set_time_usage() {
